fix: harden writer↔reader contract edges + cover with tests
Audit of the backend(writer)↔frontend(reader) data contract surfaced four
low-severity hardening gaps off the happy path. Fix each and add tests that
guard the behavior.

- db.gz first_fetched: drop omitempty (db.go) so the key is always emitted —
  the reader divides by it (data.ts findChronForTimestamp); add a reader-side
  `?? 0` guard (data.ts) and make the type optional (types.d.ts) so a
  hand-edited/migrated db.gz can't produce NaN time-jumps.
- processing.go: strip DEL (0x7f) and C1 (0x80–0x9f) controls, which survive
  #sanitize/#minify and would otherwise reach the reader.
- idx.ts: widen fetchedAts Uint16→Uint32 to remove the ~60y silent-wrap ceiling.
- store: stamp Cache-Control per object class (cacheControlForKey in main.go,
  wired into s3.go) — finalized packs immutable; db.gz + latest revalidate.

Tests: store_test cache-control classes; processing_test sanitize trust
boundary + post-#sanitize ordering hazard; db_test first_fetched always
emitted; channel_test DEL/C1 strip + NBSP preserved; idx.test Uint32 no-wrap;
pipeline e2e real-DOM no-executable-nodes assertion. `make verify` green.

+++ frontend/src/js/idx.ts
@@ -3,7 +3,7 @@ const IDX_HEADER_SIZE = 259 * 4 // 3 state uint32 + 256 chanCounts uint32
 
 export interface IdxPack {
    chanIds: Uint8Array
-   fetchedAts: Uint16Array
+   fetchedAts: Uint32Array
    chanCounts: Uint32Array
    ownChanCounts: Uint32Array
    bounds: { packId: number; startChron: number }[]
@@ -42,7 +42,7 @@ export function makeIdxPack(buf: ArrayBuffer, packIndex: number, packSize: numbe
       chanIds: new Uint8Array(0),
       chanCounts: new Uint32Array(0),
       ownChanCounts: new Uint32Array(0),
-      fetchedAts: new Uint16Array(0),
+      fetchedAts: new Uint32Array(0),
       bounds: [],
       parse() {
          if (!rawBuf) return pack
@@ -64,7 +64,10 @@ export function makeIdxPack(buf: ArrayBuffer, packIndex: number, packSize: numbe
          }
 
          const chanIds = new Uint8Array(packSize)
-         const fetchedAts = new Uint16Array(packSize)
+         // Uint32 (not Uint16): fetchedAt accumulates 8h-blocks since
+         // first_fetched; a Uint16 ceiling wraps after ~60y of continuous
+         // fetching (65535 blocks), silently corrupting time-range jumps.
+         const fetchedAts = new Uint32Array(packSize)
          pack.chanIds = chanIds
          pack.fetchedAts = fetchedAts
          const bytes = new Uint8Array(rawBuf)
